Conflicts:
	MemeCreator/bin/MemeCreator.apk
	MemeCreator/bin/classes.dex
	MemeCreator/src/cs4295/memecreator/WelcomeScreenActivity.java

+++ MemeCreator/src/cs4295/memecreator/MemeEditorActivity.java
@@ -151,9 +151,8 @@ public class MemeEditorActivity extends Activity {
 		memeEditorLayout = (LinearLayout) findViewById(R.id.memeEditorLayout);
 		memeEditorLayout.setGravity(Gravity.CENTER);
 		try {
-			Log.i("path", imagePath);
+			Log.i("imagePath", imagePath);
 			Bitmap bitmap = BitmapFactory.decodeFile(imagePath);
-			Log.i("path", bitmap.toString());
 			memeEditorView = new MemeEditorView(this, bitmap);
 			memeEditorView
 					.setLayoutParams(new LayoutParams(memeSize, memeSize));
@@ -203,24 +202,26 @@ public class MemeEditorActivity extends Activity {
 
 	// Delete a files
 	private void deleteFile(File file) {
-		Log.i("deleteFile", file.toString()
-				+ ((file.exists()) ? " is Exist." : "is not exist!!!!"));
+		if (file != null) {
+			Log.i("deleteFile", file.toString()
+					+ ((file.exists()) ? " is Exist." : "is not exist!!!!"));
 
-		// Check if the file exist
-		if (file.exists())
-			// Clear the file inside if it is a directory
-			if (file.isDirectory()) {
-				String[] children = file.list();
-				for (int i = 0; i < children.length; i++) {
-					File f = new File(file, children[i]);
-					if (f.delete())
-						Log.i("deleteFile", f.getAbsolutePath()
-								+ " is deleted....");
-					else
-						Log.i("deleteFile", f.getAbsolutePath()
-								+ " is not deleted!!!!");
+			// Check if the file exist
+			if (file.exists())
+				// Clear the file inside if it is a directory
+				if (file.isDirectory()) {
+					String[] children = file.list();
+					for (int i = 0; i < children.length; i++) {
+						File f = new File(file, children[i]);
+						if (f.delete())
+							Log.i("deleteFile", f.getAbsolutePath()
+									+ " is deleted....");
+						else
+							Log.i("deleteFile", f.getAbsolutePath()
+									+ " is not deleted!!!!");
+					}
 				}
-			}
+		}
 	}
 
 	@Override
